Rename the index-scheduler module into insta_snapshot
